DataEngineeringProject/requirements.txt
pandas
sqlalchemy
mysql-connector-python
pyyaml
requests
jupyter